services/embedding/requirements.txt
# Phase 3 embedding gRPC microservice runtime deps.
grpcio==1.80.0
grpcio-tools==1.80.0
protobuf>=6.0.0,<8.0.0
sentence-transformers>=3.0.0,<4
# torch is pinned by sentence-transformers; we leave it as a transitive
# dep to avoid version drift. CPU-only by default; override the base
# image to a cuda one for GPU inference.
torch>=2.2.0,<3
prometheus-client>=0.21,<1
